% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute the Confusion Matrix and Classification Metrics}
\usage{
compute_metrics(y_true, y_pred, threshold = 0.5)
}
\arguments{
\item{y_true}{Actual binary values (0 or 1).}

\item{y_pred}{Predicted probabilities (output of logistic regression).}

\item{threshold}{Threshold to classify predicted probabilities (default is 0.5).}
}
\value{
A list containing the confusion matrix and metrics.
}
\description{
Compute the Confusion Matrix and Classification Metrics
}
