% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{logistic_regression}
\alias{logistic_regression}
\title{Perform Logistic Regression}
\usage{
logistic_regression(X, y, lambda = 1e-05)
}
\arguments{
\item{X}{A numeric matrix of predictors, including an optional intercept column.}

\item{y}{A binary numeric vector (0/1) of responses corresponding to rows in \code{X}.}

\item{lambda}{A small constant for ridge regularization to stabilize matrix inversion 
(default is \code{1e-5}).}
}
\value{
A numeric vector of estimated coefficients.
}
\description{
Estimates coefficients for logistic regression using numerical optimization 
to maximize the log-likelihood function.
}
